[refactor] career-transition-grant page uses canonical marketing layout
Brings /programs/career-transition-grant in line with /about, /programs,
/join-us, /events: full-bleed MarketingHero, Breadcrumbs, Section-based
body, and a centred Apply CTA at the bottom.

Drops the bespoke GrantProgramHero panel-card hero (stats move to a
slim strip below the breadcrumbs), GrantPageSection wrapper, and
LandingBanner footer. Body content extracted into
components/career-transition-grant/* mirroring how /about composes
named section components. Hardcoded hex (#ECF0FF / #D7E4F5 / #2A5FA8)
and inline styles replaced with bluedot-* tokens.
